# chat-server/src/chat_server/generated/main.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2025-03-23T19:56:08+00:00

from __future__ import annotations

from typing import List, Optional, Union

from fastapi import FastAPI

from .models import (
    ChatMessage,
    Error,
    SubmitChatMessageCompareRequest,
    SubmitChatMessageRequest,
    SubmitChatMessageSelectRequest,
    Thread,
)

app = FastAPI(
    title='Simple Chat API',
    description='API for a chat application with a single endpoint',
    version='1.0.0',
    servers=[
        {
            'url': 'https://chat.chrissreesangkom.com/api',
            'description': 'Production server',
        },
        {'url': 'http://localhost:8000', 'description': 'Development server'},
    ],
)


@app.post('/thread', response_model=Thread, tags=['Chat'])
def create_thread() -> Thread:
    """
    Create a new thread
    """
    pass


@app.get('/thread', response_model=List[Thread], tags=['Chat'])
def get_threads() -> List[Thread]:
    """
    Get all threads
    """
    pass


@app.get('/thread/{thread_id}/chat', response_model=List[ChatMessage], tags=['Chat'])
def get_chat_messages(thread_id: str) -> List[ChatMessage]:
    """
    Get all chat messages
    """
    pass


@app.post(
    '/thread/{thread_id}/chat',
    response_model=ChatMessage,
    responses={'400': {'model': Error}},
    tags=['Chat'],
)
def submit_chat_message(
    thread_id: str, body: SubmitChatMessageRequest = ...
) -> Union[ChatMessage, Error]:
    """
    Submit a chat message and get a response
    """
    pass


@app.post(
    '/thread/{thread_id}/chat/compare',
    response_model=bytes,
    responses={'400': {'model': Error}},
    tags=['Chat'],
)
def submit_chat_message_compare(
    thread_id: str, body: SubmitChatMessageCompareRequest = ...
) -> Union[bytes, Error]:
    """
    Submit a chat message and get a response from another model
    """
    pass


@app.post(
    '/thread/{thread_id}/chat/compare/select',
    response_model=None,
    responses={'400': {'model': Error}},
    tags=['Chat'],
)
def submit_chat_message_select(
    thread_id: str, body: SubmitChatMessageSelectRequest = ...
) -> Optional[Error]:
    """
    Submit a chat message selection
    """
    pass